--- src/main/java/com/wayble/server/direction/entity/Elevator.java
@@ -1,26 +1,0 @@
-package com.wayble.server.direction.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "elevator")
-public class Elevator {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "latitude", columnDefinition = "DECIMAL(10,7)", nullable = false)
-    private Double latitude;
-
-    @Column(name = "longitude", columnDefinition = "DECIMAL(10,7)", nullable = false)
-    private Double longitude;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "facility_id", nullable = false)
-    private Facility facility;
-}
