--- src/main/java/com/wayble/server/direction/entity/Lift.java
@@ -1,29 +1,0 @@
-package com.wayble.server.direction.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "lift")
-public class Lift {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "latitude", columnDefinition = "DECIMAL(10,7)", nullable = false)
-    private Double latitude;
-
-    @Column(name = "longitude", columnDefinition = "DECIMAL(10,7)", nullable = false)
-    private Double longitude;
-
-    @Column(name = "exit_num", nullable = true)
-    private Integer exitNum;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "facility_id", nullable = false)
-    private Facility facility;
-}
