+++ src/main/java/com/wayble/server/direction/repository/EdgeRepository.java
@@ -1,6 +1,6 @@
 package com.wayble.server.direction.repository;
 
-import com.wayble.server.direction.entity.Edge;
+import com.wayble.server.direction.entity.transportation.Edge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EdgeRepository extends JpaRepository<Edge, Long> {
