+++ src/main/java/com/wayble/server/direction/repository/FacilityRepository.java
@@ -1,6 +1,6 @@
 package com.wayble.server.direction.repository;
 
-import com.wayble.server.direction.entity.Facility;
+import com.wayble.server.direction.entity.transportation.Facility;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FacilityRepository extends JpaRepository<Facility, Long> {
